+++ database/migrations/2026_07_18_000008_create_job_photos_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->foreignId('job_id')->constrained('jobs')->cascadeOnDelete();
             // See 2026_07_18_000007_create_jobs_table.php: users.id is a legacy signed
             // int on this app's live DB, incompatible with foreignId()'s unsigned bigint.
-            $table->unsignedBigInteger('uploaded_by');
+            $table->integer('uploaded_by');
             $table->string('path');
             $table->decimal('latitude', 10, 7)->nullable();
             $table->decimal('longitude', 10, 7)->nullable();
